refactor: moved sorting logic
